chore: split the block option action into separate files
