+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/actions/option/divider_option_action.dart
@@ -1,0 +1,20 @@
+import 'package:appflowy/workspace/presentation/widgets/pop_up_action.dart';
+import 'package:appflowy_popover/appflowy_popover.dart';
+import 'package:flutter/material.dart';
+
+class DividerOptionAction extends CustomActionCell {
+  @override
+  Widget buildWithContext(
+    BuildContext context,
+    PopoverController controller,
+    PopoverMutex? mutex,
+  ) {
+    return const Padding(
+      padding: EdgeInsets.symmetric(vertical: 4.0),
+      child: Divider(
+        height: 1.0,
+        thickness: 1.0,
+      ),
+    );
+  }
+}
