+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/actions/option/option_actions.dart
@@ -1,0 +1,74 @@
+import 'package:appflowy/generated/flowy_svgs.g.dart';
+import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:easy_localization/easy_localization.dart';
+
+export 'align_option_action.dart';
+export 'color_option_action.dart';
+export 'depth_option_action.dart';
+export 'divider_option_action.dart';
+export 'turn_into_option_action.dart';
+
+enum OptionAction {
+  delete,
+  duplicate,
+  turnInto,
+  moveUp,
+  moveDown,
+  copyLinkToBlock,
+
+  /// callout background color
+  color,
+  divider,
+  align,
+  depth;
+
+  FlowySvgData get svg {
+    switch (this) {
+      case OptionAction.delete:
+        return FlowySvgs.trash_s;
+      case OptionAction.duplicate:
+        return FlowySvgs.copy_s;
+      case OptionAction.turnInto:
+        return FlowySvgs.turninto_s;
+      case OptionAction.moveUp:
+        return const FlowySvgData('editor/move_up');
+      case OptionAction.moveDown:
+        return const FlowySvgData('editor/move_down');
+      case OptionAction.color:
+        return const FlowySvgData('editor/color');
+      case OptionAction.divider:
+        return const FlowySvgData('editor/divider');
+      case OptionAction.align:
+        return FlowySvgs.m_aa_bulleted_list_s;
+      case OptionAction.depth:
+        return FlowySvgs.tag_s;
+      case OptionAction.copyLinkToBlock:
+        return FlowySvgs.share_tab_copy_s;
+    }
+  }
+
+  String get description {
+    switch (this) {
+      case OptionAction.delete:
+        return LocaleKeys.document_plugins_optionAction_delete.tr();
+      case OptionAction.duplicate:
+        return LocaleKeys.document_plugins_optionAction_duplicate.tr();
+      case OptionAction.turnInto:
+        return LocaleKeys.document_plugins_optionAction_turnInto.tr();
+      case OptionAction.moveUp:
+        return LocaleKeys.document_plugins_optionAction_moveUp.tr();
+      case OptionAction.moveDown:
+        return LocaleKeys.document_plugins_optionAction_moveDown.tr();
+      case OptionAction.color:
+        return LocaleKeys.document_plugins_optionAction_color.tr();
+      case OptionAction.align:
+        return LocaleKeys.document_plugins_optionAction_align.tr();
+      case OptionAction.depth:
+        return LocaleKeys.document_plugins_optionAction_depth.tr();
+      case OptionAction.copyLinkToBlock:
+        return LocaleKeys.document_plugins_optionAction_copyLinkToBlock.tr();
+      case OptionAction.divider:
+        throw UnsupportedError('Divider does not have description');
+    }
+  }
+}
